requirements.txt
# asgiref==3.8.1
# aspose_slides==25.6.0
# bleach==6.2.0
# certifi==2025.4.26
# charset-normalizer==3.4.1
# colorama==0.4.6
# crispy-bootstrap4==2024.10
# dj-database-url==1.0.0
# # Django==5.2
# Django==5.1.4
# PyGithub==2.3.0
# django-allauth==65.5.0
# django-axes==7.0.2
# django-cleanup==9.0.0
# django-cors-headers==4.7.0
# django-crispy-forms==2.4
# django-debug-toolbar==5.1.0
# django-environ==0.12.0
# django-formtools==2.5.1
# django-honeypot==1.2.1
# django-otp==1.5.4
# django-password-validators==1.7.3
# django-phonenumber-field==8.0.0
# django-ratelimit==4.1.0
# django-two-factor-auth==1.17.0
# django_csp==3.8
# djangorestframework==3.15.2
# et_xmlfile==2.0.0
# idna==3.10
# iniconfig==2.1.0
# lxml==5.4.0
# mpmath==1.3.0
# numpy==2.2.5
# openpyxl==3.1.5
# packaging==24.2
# pillow==11.2.1
# pluggy==1.5.0
# psycopg2-binary==2.9.10
# pyotp==2.9.0
# pypng==0.20220715.0
# pytest==8.3.5
# python-decouple==3.8
# python-dotenv==1.0.1
# python-pptx==1.0.2
# qrcode==7.4.2
# requests==2.32.3
# sqlparse==0.5.3
# sympy==1.13.3
# typing_extensions==4.12.2
# tzdata==2025.2
# urllib3==2.4.0
# webencodings==0.5.1
# whitenoise==6.9.0
# XlsxWriter==3.2.3

asgiref==3.8.1
aspose-cells==24.10.0
aspose_slides==25.6.0
beautifulsoup4==4.13.4
bleach==6.2.0
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.1
colorama==0.4.6
contourpy==1.3.2
crispy-bootstrap4==2024.10
cryptography==45.0.5
cycler==0.12.1
Deprecated==1.2.18
dj-database-url==1.0.0
# Django==5.2
Django==5.1.4
django-allauth==65.5.0
django-axes==7.0.2
django-cleanup==9.0.0
django-cors-headers==4.7.0
django-crispy-forms==2.4
django-debug-toolbar==5.1.0
django-environ==0.12.0
django-formtools==2.5.1
django-honeypot==1.2.1
django-otp==1.5.4
django-password-validators==1.7.3
django-phonenumber-field==8.0.0
django-ratelimit==4.1.0
django-two-factor-auth==1.17.0
django_csp==3.8
djangorestframework==3.15.2
docopt==0.6.2
et_xmlfile==2.0.0
fonttools==4.59.0
idna==3.10
iniconfig==2.1.0
jpype1==1.6.0
kiwisolver==1.4.8
lxml==5.4.0
matplotlib==3.10.3
mpmath==1.3.0
next==1.0.1
num2words==0.5.14
numpy==2.2.5
opencv-contrib-python==4.12.0.88
openpyxl==3.1.5
packaging==24.2
pillow==11.2.1
pluggy==1.5.0
psycopg2-binary==2.9.10
pycparser==2.22
PyGithub==2.6.1
PyJWT==2.10.1
PyNaCl==1.5.0
pyotp==2.9.0
pyparsing==3.2.3
pypng==0.20220715.0
pytest==8.3.5
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
python-pptx==1.0.2
qrcode==7.4.2
requests==2.32.3
six==1.17.0
soupsieve==2.7
sqlparse==0.5.3
sympy==1.13.3
typing_extensions==4.12.2
tzdata==2025.2
urllib3==2.4.0
webencodings==0.5.1
whitenoise==6.9.0
wrapt==1.17.2
XlsxWriter==3.2.3
gunicorn==20.1.0